requirements.txt
asyncio
socketio
python-socketio[asyncio_client]
LZString
python-dotenv
